When the user click on the analyze button no indication that the symbols window is updating This update show the user that the process is working
